Added TrackInfo struct, TrackBuffer typedef.
The VolumeBuffer typedef wasn't able to uniquely identify the VolumeInfo
objects, since they are unique on Track ID, SampleID, ClientGUID.

VolumeBuffer is now a map from sampleid_t to maps from ClientGUIDs to
VolumeInfo objects.

Additionally, each set of Buffer objects are maintained within a
TrackInfo object which alleviates the requirement of having the TrackID
be part of the identification key.