Implemented receiveClientConnections.
Added the bool operator< () to the ClientGUID structure to make it
std::map compatible. To do so, I combined the BroadcastIP and LocalIP
into an unsigned __int 64 using a union. Whether or not comparing
ClientGUID's like that makes any sense, the map will work as a
consequence of the consistency. Why the hell std::map needs the <
operator and not equality is beyond me.

Even if it did two < operators to make sure two things were equal, it
would be incredibly stupid.

Anyway, getMicroseconds() was also implemented inside of
NetworkGlobals.cpp to facilitate receiveClientConnections.