Began implementation of Request::PLAY; Reworked buffering mechanism.
Added mechanisms to ensure delivery of a range of samples. Updated The
Megadoc under 3.2.8 Send Audio Data.

Having the clients determine if they're missing a packet without knowing
what they're supposed to be getting, only what they need to be playing
in the future seems difficult to implement and shaky in any event, and
having clients send acknowledgements for each packet would flood the
network. sendAudioSample now requires an intended buffering range to be
sent along with its sample packets. The clients will know if they're
missing part of the buffer range if they do not receive the full range
by the end of a certain timeout, in which case a request will be sent
out for a resend.

Added a global variable for the number of packets required to be
buffered before PLAY transport control packets are dispersed.

Added a global variable for the number of packets that are buffered on a
regular BUFFER request.
- BUFFER requests begin immediately after initial buffering is complete,
are queued up at the end of each successful BUFFER request, and are
limited by the amount of samples they buffer. Other server requests can
interleave BUFFER requests, allowing the server to continue being
functional while buffering its clients.