Created the PacketStructures::NetworkMessage structure/paradigm.
So reading and writing messages with pointer type casting and arithmetic
was really ugly, non-uniform, and frankly probably dangerous. To fix
this, Networking::PacketStructures::NetworkMessage was introduced.

NetworkMessage is a 32 byte message header struct that contains all of
the information regarding a message sent on our network. To send, or
when receiving, a packet on our network, you should ALWAYS use a
NetworkMessage header.

Implementation Details:
format: [ byte count ] VariableName - Usage

[1] ControlByte - Indicates the purpose of this packet
[7] _pad - Explicit byte padding.

[16] Union of Headers - Header information for every kind of network
message that needs to be sent

[8] data - Set to 8 for padding purposes, but some packets carry extra
information, and may be allocated larger than 32 bytes. If that's the
case, then the extra information (and the first 8 bytes of which) start
at data.

--

Padding is explicitly introduced to make the structure compatible across
compiler standards. MSVC++ aligns structures to an 8 byte boundary. If
g++ would align differently for the pi's, it won't make the structure
incompatible.

Class layout from MSVC:

2>  class NetworkMessage size(32):
2>   +---
2>       | <unnamed-tag> <alignment member> (size=0)
2>   0  | <unnamed-type-ClientConnection> ClientConnection
2>   0  | <unnamed-type-ClientCheckIn> ClientCheckIn
2>   0  | <unnamed-type-AudioSample> AudioSample
2>   0  | <unnamed-type-AudioResendRequest> AudioResendRequest
2>   0  | <unnamed-type-VolumeSample> VolumeSample
2>   0  | <unnamed-type-VolumeResendRequest> VolumeResendRequest
2>       | <unnamed-tag> <alignment member> (size=0)
2>  16 | ControlByte
2>  17 | _pad
2>  24 | data
2>   +---