Slight update to PacketStructures::NetworkMessage.
The last 8 bytes (PacketStructures::NetworkMessage::Extra) are no longer
to be considered the first 8 bytes of extra data. The structure is 32
bytes long, and our packets can be at most 1500 bytes long. Therefore
any extra data must always fit int 1500 - 32 = 1468 bytes.

This extra data must start at
(PacketStructures::NetworkMessage::Extra::_data + 8). The last 8 bytes
of the structure can be left unused (in which case, they are to be
considered alignment padding), or can be used in whatever fashion
necessary. For convenience, if the length of the data needs to be
transmitted in addition to the data, the union
PacketStructures::NetworkMessage::Extra defines another member,
_dataLength as a size_t for that purpose.