Added client dropping + safeguards in CheckIn.
Client dropping required a global constant indicating the number of
milliseconds between check-ins to remain "active". CLIENT_CHECKIN_DELAY
was defined in NetworkConstants.h and set to 300ms.

Clients are dropped on the TIMER_TICK request processing (every 200ms)
if they have been inactive for a total of CLIENT_CHECKIN_DELAY ms. In
theory, a client could be inactive for a total of 200ms (TIMER_TICK
delay) + CLIENT_CHECKIN_DELAY ms, but surely this isn't a serious issue.

Safeguards were added to receiveClientCheckIn. If a client checks-in
that has been dropped due to inactivity, they are no longer considered
in this function.