Added the NetworkSpecs.txt.
NetworkSpecs.txt provides a full detailed description of how the Client
should expect to receive messages and how the Server WILL be expecting
to receive messages from the Client.

Further, it provides a possible algorithmic solution to detecting
dropped packets in the network. The solution is based on the sample ID
range information sent along with audio samples and volume data from the
server.