Fixed possible transport control request bug.
Previously, the situation was possible where a necessary transport
control like PAUSE or STOP would be initiated. If not every single
client responds to one of these transport control messages, the server
continues to send them out until they all have responded (meaning that
they have all paused or stopped). If one of the clients times out during
at the time of a necessary transport control request, it is highly
unlikely that they will ever send an acknowledgement for the request,
and the server will hang waiting for it.

Now, the server's client dropout system will allow a client to be
dropped out of the request acknowledgement system and cease to affect
whether or not the server can continue processing requests.