Guard fixes, log fixes, IP fixes
Fixed the PI guards so that compilation works on both systems. Pi
compilation now requires a -DRASPBERRY_PI flag.
Fixed converting the IP addresses from strings to client IDs in the
client software.
Fixed logging client ID in the server software to print a string,
instead of trying to print an integer.