Bug fixes to server, tested with small test client
1. Fixed a bug in the sockets.h/sockets.cpp that made it not work in
windows. - Apparently you can't use regular file descriptor read/write
with socket descriptors in Windows.
2. Likewise, there's a closesocket function in windows
3. Fixed a bug where the server would spawn up new server sockets for
every connection, instead of just using the same server socket and
passing off accepting clients.
4. Took out OnClientCheckin callbacks in the FFI.
5. Added a function to the sockets.h/sockets.cpp to "receive until
exactly this many bytes have been processed", since using TCP apparently
means that we can receive 48 bytes in a packet when we asked for 1468.
6. Fixed a bug where the log file wasn't being opened in time for the
logging functions and causing race conditions on the file pointer.
7. Fixed a bug where audio samples weren't being buffered past initial
buffering because the condition for re-buffer was the same as in the
initial buffering state.
8. Fixed a bug where the final audio buffer range was way, way less than
the ID of the last sample that needed to be buffered.